Allow label to be specified in release asset upload
Update UploadOptions with a Label field. The label is provided as a URI
query parameter to the GitHub API request when uploading the file. The
label is used as a short or alternate display name for the file in the
release.

Update unit tests to test for two scenarios, the label being provided
in UploadOptions and the label being excluded. When the label is
provided it should be a query param in the request.