Use embeded CheckSuite and HeadCommit field for CheckSuiteEvent
The `check_suite` object from github's check_suite event also includes `head_commit`, which is not present in the `check_suite` object.

The `CheckSuiteEvent` also uses `CheckSuite` as one of its fields in the struct which does not has the head_commit, and therefore is not parsed during json parsing in the current `CheckSuiteEvent`, hence `head_commit` data is unavailable from the `CheckSuiteEvent` struct.

To also parse the `head_commit`, `CheckSuiteEvent` should use an embeded field for `CheckSuite` with `HeadCommit`, so that all the details are received from the github api.

```go
type CheckSuiteEvent struct {
	CheckSuite struct {
		*CheckSuite
		HeadCommit *Commit `json:"head_commit,omitempty"`
	} `json:"check_suite,omitempty"`
	// The action performed. Can be "completed", "requested" or "re-requested".
	Action *string `json:"action,omitempty"`

	// The following fields are only populated by Webhook events.
	Repo         *Repository   `json:"repository,omitempty"`
	Org          *Organization `json:"organization,omitempty"`
	Sender       *User         `json:"sender,omitempty"`
	Installation *Installation `json:"installation,omitempty"`
}
```

This would remove a function from `github-accessors.go`

```go
// GetCheckSuite returns the CheckSuite field.
func (c *CheckSuiteEvent) GetCheckSuite() *CheckSuite {
       if c == nil {
               return nil
       }
       return c.CheckSuite
}
```