scrape: support textarea, radio, and checkbox
improve form parsing to support textarea elements, as well as proper
handling of radio and checkbox inputs (only include the value if the
"checked" attribute is set).

also allow nil setValue func for fetchAndSubmitForm.  While I don't know
of a case (yet) where we want to do this, there's no point in panicking
in this case.