Added a new Example using libsodium to create a new secret in a github repo for use in github actions.
It is usable as a cli package to produce a secret in a given repo that the user owns.

usage is
main creates a new secret in github for a given owner/repo based on a secretName passed in as an arg
and a github token, and secret value provided via the environment

usage:
	export GITHUB_AUTH_TOKEN=<auth token from github that has secret create rights>
	export SECRET_VARIABLE=<secret value of the secret variable>
	go run main.go -owner <owner name> -repo <repository name> SECRET_VARIABLE

ex:
	export GITHUB_AUTH_TOKEN=0000000000000000
	export SECRET_VARIABLE="my-secret"
	go run main.go -owner google -repo go-github SECRET_VARIABLE

Once it runs go to the github repository > settings > left side options bar > Secrets
And you should see the new secret appear there.