Add Support For Millisecond Timestamps
The GitHub Audit Log API introduced timestamps
with millisecond granularity. We add support
for these timestamps by first parsing the timestamp
into a Unix timestamp and then checking to see if
the generated timestamp is in the future. If the
timestamp is in the future we re-parse it as a Unix
timestamp in Milliseconds.

Signed-off-by: Brett Logan <lindluni@github.com>