editChange event: define struct types instead of using anonymous structs
The types of the fields of the EditChange struct were defined inline as
anonymous structs.
This made it tedious to use the EditChange struct in testcases, because
the anonymous type needed to be redefined (incl the exact same JSON
tags) whenever it was assigned. The TestEdit* testcases are good example
for it.

Using anonymous struct also does not scale well when the struct changes.
For example when a field is added to an anonymous struct, every place
where the anonymous struct is assigned must also be changed.

Another advantage of using regular types is that accessors  will be
generated automatically and can be used instead of having to manually
check for nil values when accessing the struct fields.

Define the types as regular structs in event_types.go instead.