Add allows_public_repositories to actions_runner_group
Adds allows_public_repositories to request bodies of the
CreateRunnerGroup as well as UpdateRunnerGroup methods for github
actions.

This allows us to set the allows_public_repositories flag even if it is
not yet properly documented see https://github.com/github/docs/issues/8209
That flag controls if public repos can use runners of this group.
If it is disabled(default) even public repos that are added to the
selected_repositories list cannot use runners of this group.

With this change we can now set this properly and thus use runner groups
created by go-github also for public repos.

Fixes #1996