Expose the package version on the `Client` (e.g. `"45.2.0")
This exposes the current version of the `go-github` package
(e.g. `"45.2.0"`) as a field on the `Client` struct.

This will allow users of `go-github` to access the current version
of the package, for example if they want to construct a custom
`User-Agent` header including it.

This change means that, before releasing tagging and releasing a
new version of the package, this string constant will have to be
updated.