Include the version of the module in the default `User-Agent` request header
This updates the default request header sent by the module to
include the current version, e.g. `go-github/45.2.0`.

This makes it easier for GitHub to track usage of this SDK and
how users are upgrading between versions.

Users can continue to overwrite this default user agent with
their own string by setting the `UserAgent` property on the
`Client`.