Remove GenerateReleaseNotes from RepositoriesService.EditRelease
According to the [GitHub v3 API reference](https://docs.github.com/en/rest/releases/releases#update-a-release),
`generate_release_notes` is not a recognised body parameter of the endpoint to update a release. If provided to
the API, it will have no effect.

This commit removes the setting of this parameter when updating a release, to avoid confusion in the effect of
setting this parameter.