Add MakeLatest parameter
GitHub recently changed the way the latest version can be managed, to
allow explicitly setting the latest release. Previously, the latest
release would be the most recently created non-pre-release.

https://github.blog/changelog/2022-10-21-explicitly-set-the-latest-release/

In the web UI, this shows up as two checkboxes. In the API, this is set
using the make_latest parameter, which defaults to true for newly
created releases.

https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28#update-a-release

https://github.com/google/go-github/issues/2593