fix: relax go directive in go.mod to 1.22.0
Rather than require all consumers to be building with the latest 1.22.10
release of Go 1.22, match the support statement and permit any version
of 1.22.x to be used. Since Go 1.21 the go directive in go.mod is
considered a hard requirement version and anything older will refuse to
build.

Update the N-1 go-version used in GitHub Actions to be the .0 release to
help ensure this compatibility is maintained in the future.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>