Add support for registry_package event
as of the time of this writing, the Github documentation describes two
events related to packages. One called package and one called
registry_package. According to the documentation, users should switch to
using the package event instead of the registry_package event. However,
when working with the actual permissions and events, there is no package
event but there is a RegistryPackage event that can be enabled for a
Github App. It looks like the documentation is wrong and the
registry_package event was added after the package event and it's the
registry_package event that users should expect to receive.